+++ internal/pkg/managementcluster/types.go
@@ -1,0 +1,8 @@
+package managementcluster
+
+type ManagementCluster struct {
+	Name      string
+	Namespace string
+	Provider  string
+	Region    string
+}